ArraySet: Fix constructors and add assignment operators
Because Begin is const there weren't assignment operators, but we can
implement them, because Begin always points to the begin of our Data.

Also the copy and move constrcutor are not allowed to copy Begin and End,
but have to adapt them to our local Data.

No destructor implemented or defaulted, because we do not need that and it
would hinder the constexpr usage of ArraySet, which was the reason to
implement them.